Extract Request and Result classes from NetSuite::Client
NetSuite::Client is getting a bit large and with retrieving fields
from NetSuite and having to build yet another object, it makes sense to
extract these now.

* In support of PR #162
* For: https://trello.com/c/mThbJz4M/
